Refactoring pastas e tipos classes
