+++ src/main/java/network/entidade/ComunicadoDeDesligamento.java
@@ -1,6 +1,8 @@
 package network.entidade;
 
+import network.entidade.base.Comunicado;
+
 /**
  * Classe para comunicar desligamento entre cliente e servidor
  */
-public class ComunicadoDeDesligamento extends Comunicado {}
+public class ComunicadoDeDesligamento implements Comunicado {}
