+++ src/main/java/network/entidade/base/Comunicado.java
@@ -1,8 +1,8 @@
-package network.entidade;
+package network.entidade.base;
 
 import java.io.Serializable;
 
 /**
  * Classe base para comunicação entre cliente e servidor
  */
-public class Comunicado implements Serializable, Cloneable {}
+public interface Comunicado extends Serializable, Cloneable {}
